backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
tensorflow
pillow
numpy
opencv-python
